+++ deprecated/settings.py
@@ -1,0 +1,7 @@
+BLUE = (0,0,255)
+BLACK = (0,0,0)
+RED = (255,0,0)
+YELLOW = (255,255,0)
+
+SQUARESIZE = 100
+RADIUS = int(SQUARESIZE/2 - 5)
