+++ utils.py
@@ -11,9 +11,9 @@ import random
 
 
 def create_file(name):
-    if not os.path.exists("data/"):
-        os.mkdir("data")
-    file = open(f"data/{name}.csv", "w", newline="")
+    if not os.path.exists("plot_data/"):
+        os.mkdir("plot_data")
+    file = open(f"plot_data/{name}.csv", "w", newline="")
     return file
